+++ lib/core/_store.dart
@@ -1,7 +1,7 @@
 import 'package:meowchannel/core/dispatcher.dart';
 
 abstract class AbstractStore<S> extends StoreDispatcher {
-  void close();
+  Future<void> close();
   Future<S> getState();
   S getPreviousStateUnsafe();
   S getStateUnsafe();