+++ lib/core/reducer.dart
@@ -1,3 +1,3 @@
 import 'package:meowchannel/core/action.dart';
 
-typedef Reducer<S> = S Function(Action action, S previousState);
+typedef Reducer<S> = Future<S> Function(Action action, S previousState);