+++ lib/extensions/cached_worker/cache_state.dart
@@ -1,6 +1,7 @@
 
 import 'package:dataclass/dataclass.dart';
 import 'package:meowchannel/extensions/combined_reducer.dart';
+import 'package:meowchannel/extensions/synced_reducer.dart';
 import 'package:meowchannel/extensions/typed_reducer.dart';
 
 import '_cache_action.dart';
@@ -17,18 +18,18 @@ class CacheState extends _$CacheState {
 }
 
 final cacheReducer = combinedReducer<CacheState>([
-  typedReducer<SetCacheAction, CacheState>(
+  typedReducer<SetCacheAction, CacheState>(syncedReducer<SetCacheAction, CacheState>(
     (action, previousState) => 
       previousState.copyWith(
         caches: Map.fromIterable(previousState.caches.entries.toList() + [MapEntry(action.key, action.validUntil)])
       ), 
       initialState: CacheState()
-  ),
-  typedReducer<RemoveCacheAction, CacheState>(
+  )),
+  typedReducer<RemoveCacheAction, CacheState>(syncedReducer<RemoveCacheAction, CacheState>(
     (action, previousState) => 
       previousState.copyWith(
         caches: Map.fromIterable(previousState.caches.entries.where((entry) => entry.key != action.key))
       ), 
       initialState: CacheState()
-  ),
+  )),
 ]);