+++ lib/extensions/channel.dart
@@ -85,8 +85,8 @@ class StateChannel<T> {
 }
 
 //
-//Like really? 
-//This lang does not have stack implementation?!
+// Like really? 
+// This lang does not have stack implementation?!
 //
 class _Stack<T> {
   final int cachedItems;