+++ lib/extensions/combined_reducer.dart
@@ -3,9 +3,11 @@ import 'package:meowchannel/core/reducer.dart';
 
 Reducer<S> combinedReducer<S>(
   Iterable<Reducer<S>> reducers
-) => (Action action, S previousState) {
-  return reducers.fold<S>(
-    previousState, 
-    (previousState, reducer) => reducer(action, previousState)
+) => (Action action, S previousState) async {
+  return reducers.fold<Future<S>>(
+    Future.sync(() => previousState), 
+    (previousState, reducer) async {
+      return reducer(action, await previousState);
+    } 
   );
 };