+++ test/2_watcher/values_reducer.dart
@@ -4,7 +4,7 @@ import 'values_actions.dart';
 import 'values_state.dart';
 
 final Reducer<ValuesState> valuesReducer =
-  typedReducer<ValuesAction, ValuesState>((ValuesAction action, ValuesState previousState) {
+  typedReducer<ValuesAction, ValuesState>(syncedReducer((ValuesAction action, ValuesState previousState) {
     //Does nothing
     return previousState;
-  });
+  }));