+++ test/3_ui/note_list_reducer.dart
@@ -4,9 +4,9 @@ import 'note_list_actions.dart';
 import 'note_list_state.dart';
 
 final Reducer<NoteListState> noteListReducer =
-  typedReducer((Action action, NoteListState previousState) {
+  typedReducer(syncedReducer((Action action, NoteListState previousState) {
     return previousState.copyWith(
       noteList: action is NoteListUpdateStateAction? action.noteList
         : previousState.noteList
     );
-  });
+  }));