+++ test/5_multi_provider/root_reducer.dart
@@ -6,19 +6,19 @@ import 'root_state.dart';
 
 
 final Reducer<FirstRootState> firstRootReducer =
-  typedReducer<FirstRootAction, FirstRootState>((FirstRootAction action, FirstRootState previousState) {    
+  typedReducer<FirstRootAction, FirstRootState>(syncedReducer((FirstRootAction action, FirstRootState previousState) {    
     return previousState.copyWith(
       value: action is FirstRootIncreaseAction? previousState.value + 1 
         : action is FirstRootDecreaseAction? previousState.value - 1 
         : previousState.value
     );
-  });
+  }));
 
 final Reducer<SecondRootState> secondRootReducer =
-  typedReducer<SecondRootAction, SecondRootState>((SecondRootAction action, SecondRootState previousState) {    
+  typedReducer<SecondRootAction, SecondRootState>(syncedReducer((SecondRootAction action, SecondRootState previousState) {    
     return previousState.copyWith(
       value: action is SecondRootIncreaseAction? previousState.value + 1 
         : action is SecondRootDecreaseAction? previousState.value - 1 
         : previousState.value
     );
-  });
+  }));